Rework and refactor FourChanAPIE so it's safer
I noticed we were losing content that had already been downloaded
everytime we KeyboardInterrupted archive-chan or it errored out
because we were rendering and saving the threads only after downloading
every media file on it.

Since downloading media files is the slowest part of the process,
I decided to save each thread twice.

The first time quickly right after grabbing the JSON from 4chan's
API and then get to work on downloading the media.
After downloading the media, overwrite with the appropriate data.